requirements.txt
asn1crypto==0.24.0
cffi==1.11.5
Click==7.0
cryptography==2.4.2
enum34==1.1.6
Flask==1.0.2
Flask-HTTPAuth==3.2.4
Flask-JSONRPC==0.3.1
idna==2.8
ipaddress==1.0.22
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
pycparser==2.19
pyOpenSSL==18.0.0
six==1.12.0
Werkzeug==0.14.1
